wil6210: slave mode, platform device infrastructure
1. Use platform device for interaction with master driver.
This removes the module dependency between slave and master
drivers and allows loading at any order.
2. Updated slave operations. Register/unregister master is
now part of the ops. The ops are passed to master driver
using platform data.
3. connected rop - slave now assigns and reports 2 link IDs
per connection, for TX and RX. Currently we assign the
link IDs in the driver but later they will be provided
by FW. Main purpose of link IDs is to assign different
network properties to each connection such as a separate
network queue. Ths slave driver passes the link IDs
as-is to master driver.

Change-Id: Ic94520a93807118bc6d1c187bf2b4afcfd64a9b2
Signed-off-by: Lior David <liord@codeaurora.org>