wil6210: use wigig.ini as a configuration file
Adding a configuration file wigig.ini to enable uniform cross platform
configuration. Wil6210 maintains configuration table, with default
configuration values that can be overwritten with values from
wigig.ini file.
Wil6210 will read the wigig.ini file and parse it. It will check all
parameters defined in wil6210 configuration table, in case they appear
in the ini file, it will override it with the new value.

The format of the ini file is as follows:
1- Comment line as below
2- Configuration line as below
param1=value1

The ini parser will check validity of the parameters and will check
if its values are in the [min,max] range, and if the value exceeds one
of these values the parser will override with the [min,max] limit.

Some of these parameters need a unique handler function that knows
the parameter functionality, its input format and its number of input
values and how to store it.
The handler function reads, parses and stores the values related to
the new parameter into the driver structures. If needed the driver sends
the relevant FW command when the interface comes up.

Change-Id: If98c4ef780a4653d1dbc9dc9f8dae9b52a98527d
Signed-off-by: Ahmad Masri <amasri@codeaurora.org>