Allow wil_fw_dump to read log from PCIe devices directly
Add --device/-d argument that accepts PCI-e device selector in sthea
full 0001:01:00.0-like form. When invoked with -d, the tool will mmap
the PCIe bar and will fetch the log data from the device directly.

When used in that mode, the tool resets the log levels for all modules
in the firmware to print errors only, with the exception of FW_3P,
a.k.a. Module15. FW_3P module id and Info log level are used to output
logs produced by the Facebook firmware.

Signed-off-by: Alexander Kabaev <kan@fb.com>
Signed-off-by: Frank Li <frankli1@fb.com>
Signed-off-by: Michael Callahan <michaelcallahan@fb.com>