wil6210-dpdk: implement nl60g API
Implement the same nl60g API from the wil6210
kernel driver. This is needed for tools such
as host_manager_11ad to work with the DPDK driver.

The implementation uses a local socket instead of
the netlink socket, but the OTA protocol is
identical, in order for tools to reuse most of
their existing implementation, as well as make
it easy to port new nl60g API when added to
wil6210 kernel driver.

Currently supports the existing nl60g API from
the driver including getting driver capabilities
and FW state, sending WMI commands and receiving
WMI events. The DPDK driver implementation did
not support the FW state monitoring so this was
also added based on wil6210 implementation.

The initial change does not support read/write
of memory registers/blocks, as this is implemented
in the wil6210 driver using ioctl and not
through nl60g API. Support will be added in a
future change.

Change-Id: Ic8a0cd2abe2c9ecc3254875eb1a1b1b7ced87ed4
Signed-off-by: Lior David <liord@codeaurora.org>