wil6210-dpdk: add no_fw_recovery and mtu_max devargs
The no_fw_recovery and mtu_max were module
parameters in the wil6210 kernel driver. Add them
as devargs since they are also needed for VPP/DPDK.
Also store their values inside the wil6210_priv
structure so it will be possible to configure
different values for each device (RF).
Note, in the wil6210 driver no_fw_recovery could
be modified in runtime but it is load-time
argument here, as there is no real need to
change it at runtime.

Change-Id: If9774d4f8fedf7c35d0e1d6906d1066f03386a6f
Signed-off-by: Lior David <liord@codeaurora.org>