wil6210-dpdk: add GET_STA_INFO nl60g API
Added nl60g generic command GET_STA_INFO which returns
basic information about all stations in the driver
stations table. This provides the same information
as the debugfs "stations" file but without the
statistics, crypto and rx reorder information.
Returns information as an array where the index in
the array is the CID. For each CID, returns its
connection status and if connected - MAC address,
MID and AID.

Change-Id: If6bb3dac4f7a98631cda30ca9e7f49a4b3ee41bd
Signed-off-by: Lior David <liord@codeaurora.org>