wil6210-dpdk: support legacy PMC using nl60g API
Add support for legacy PMC using nl60g API.
It provides the same functionality as the
wil6210 driver debugfs files: pmccfg, pmcdata,
pmcring.
The following commands are supported:
- ALLOC: allocate PMC descriptors. Number of
descriptors and payload size for each descriptor
can be specified.
- FREE: free the PMC ring.
- GET_DATA, GET_DESC_DATA: read the payload from
PMC descriptors, or the descriptors themselves.
Caller can specify 0 bytes to read in order to
receive the amount of bytes available for reading.
The data can be read in multiple calls, and the
reply will contain the MORE_DATA flag if more data
is available for reading.

Change-Id: I34ef51c208863760a2120df8fc0a4cd207cd6b00
Signed-off-by: Lior David <liord@codeaurora.org>