wil6210-dpdk: initialize userdata in allocated RX skb
The skb->userdata is currently cleared before each
RX skb is returned to the pool, but only for the
first skb in a fragment chain. If there are fragmented
packets, there is a risk some of the skbs will have
uninitialized userdata field. Fix this by always
initializing userdata when allocating the skb.

Change-Id: Ie9c8a2c0aeca8c17b90daffcfaa9b3c0e4b68822
Signed-off-by: Lior David <liord@codeaurora.org>