wil6210-dpdk: remove arrival rate calculations
Remove arrival rate calculations in the PMD that are just based off of
accumulated bytes pending. Instead use whatever arrival rate is passed
in.

This is done because the bytes pending that get seen by the PMD do not
include the bytes from packets that are dropped by the hqos scheduled
in VPP, so the arrival rate calculated here is not the true arrival
rate. The true arrival rate that includes packets being dropped will
be calculated in VPP and sent down.

Signed-off-by: Frank Li <frankli1@fb.com>
Signed-off-by: Michael Callahan <michaelcallahan@fb.com>