wil6210-dpdk: remove arrival rate calculations
Remove arrival rate calculations in the PMD that are just based off of
accumulated bytes pending. Instead use whatever arrival rate is passed
in.

This is done because the bytes pending that get seen by the PMD do not
include the bytes from packets that are dropped by the hqos scheduled
in VPP, so the arrival rate calculated here is not the true arrival
rate. The true arrival rate that includes packets being dropped will
be calculated in VPP and sent down.

Change-Id: I1165dda9fb4c488c0825dbd43dbcb3986b43538c
Signed-off-by: Frank Li <frankli1@fb.com>
Signed-off-by: Michael Callahan <michaelcallahan@fb.com>
(cherry picked from commit 1b9186048f8da2232f78bb344ff0ee22f0755a36)
Signed-off-by: Lior David <liord@codeaurora.org>