wil6210-dpdk: set nl60g poll worker affinity
The nl60g API was added by Qualcomm to the PMD to work with tools like
host_manager_11ad for things like getting driver capabilities, getting
FW state, sending WMI commands etc.  The poll worker thread to manage
connections does not have a CPU affinity and is marked with a TODO
question for it.

We want to ensure that the nl60g poll worker only runs on the VPP
master lcore, and other cores are properly reserved.

Signed-off-by: Frank Li <frankli1@fb.com>
Signed-off-by: Michael Callahan <michaelcallahan@fb.com>