wil6210-dpdk: support getting FW capabilities using nl60g API
Added a generic command NL_60G_GEN_GET_FW_CAPA that
will return the FW capabilities. It is returned as
variable size array containing one or more 32 bit
units in little endian format.
The FW capabilities can be used to detect various
FW capabilities. For example, the FW may support
some special WMI but if trying to call this WMI
on a FW that does not support it, the FW will
shut down. By using the FW capabilities it is
possible to know in advance whether it is ok to
call the WMI on this specific FW.

Change-Id: Ic458c258909a49d82f952b4bae57b91fbddd600a