wigig-dpdk: support recovery of downgraded PCIe link
Add support for recovery of downgraded PCIe link for
Talyn card.
This is done by passing the expected PCIe link
configuration (speed and lanes) to the DPDK driver
as devargs.
When device is started, the current PCIe link
configuration will be compared to the expected one.
If they are not the same, attempt to recover by
retraining the link.

Change-Id: I4532597c0baf45061b8d87a30af36c7bf547c5e2