wil6210: support recovery of downgraded PCIe link
Add support for recovery of downgraded PCIe link for
Talyn card.
This is done by setting the expected PCIe link
configuration (speed and lanes) to the wil6210
driver using sysfs.
When device is started, the current PCIe link
configuration will be compared to the expected one.
If they are not the same, attempt to recover by
retraining the link.

Change-Id: Icb4d8046d233cab96720341a7d9b4b25e30dee36
Signed-off-by: Lior David <liord@codeaurora.org>