wigig-dpdk: optimize PMC descriptors allocation
The existing code allocates a separate chunk of memory
for each PMC descriptor. This can take more time for
larger number of descriptors. For example, when using
1024 descriptors the host tools may fail starting PMC
due to timeout waiting for PMC allocate call to return
(mainly happens with multiple BBs).
In order to fix this, allocate one large chunk for
all descriptors with size num_descriptors * descriptor_size.
Then each descriptor memory is initialized as offset to
this memory. This way allocation will be much faster,
and less sensitive to number of descriptors.

Change-Id: Ic4eb8c1d647ed7207c040227650d48ae11b59329