wil6210: crash dump support for PMC ext host mode
FW logs used to be in FW device memory, on crash dump, wil6210 used to
copy all FW device memory sections, and it included the FW logs

By moving to PMC continuous mode, in host mode, the PMC data is allocated
on host memory, and when wil6210 dumps the devices memory, the user may
miss the FW logs.

In this patch wil6210 is dumping the PMC data to the same buffer it
dumps the fw, using some tools the user can retrieve the latest logs
from the FW on crash.

Change-Id: I319ffd4502290a4b966e0025522b10ff2a23bfef
Signed-off-by: Ahmad Masri <amasri@codeaurora.org>
Signed-off-by: Visweswara Tanuku <vtanuku@codeaurora.org>